Renaming an atavism
treeParser was introduced while there was still a parser which did
things with pegase's ASt. It was never renamed after removing the
pegase related AST code. This has been rectified.
