+++ client/src/features/Parse/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { treeParser, Scope } from "../../common/treeParser";
+import { parser, Scope } from "../../common/parser";
 import { Expression } from "../Expression";
 import styles from './Parse.module.css'
 
@@ -11,7 +11,7 @@ export type ParseProps = {
 export const Parse = (props: ParseProps) => {
   console.log('parsing expression:', props.input)
   try {
-    const output = treeParser.value(props.input, {context: props.scope})
+    const output = parser.value(props.input, {context: props.scope})
     return <div className={styles.result}>
       <span>{'=>'}</span>
       <Expression node={output} />
